more extraction and mock drawio graph get model
