+++ src/nosql-ts.ts
@@ -1,5 +1,5 @@
 import { DbDefinition, DbRelationshipDefinition } from "@funktechno/little-mermaid-2-the-sql/lib/src/types";
-import { TableAttribute, TableEntity } from "./types/sql-plugin-types";
+import { DatabaseModelResult, TableAttribute, TableEntity } from "./types/sql-plugin-types";
 import { DatabaseModel, ForeignKeyModel, PrimaryKeyModel, PropertyModel, TableModel } from "@funktechno/sqlsimpleparser/lib/types";
 import { JSONSchema4, JSONSchema4TypeName } from "json-schema";
 import { convertCoreTypesToJsonSchema, convertOpenApiToCoreTypes, jsonSchemaDocumentToOpenApi } from "core-types-json-schema";
@@ -8,6 +8,7 @@ import { convertTypeScriptToCoreTypes } from "core-types-ts/dist/lib/ts-to-core-
 import { convertCoreTypesToTypeScript } from "core-types-ts";
 import { GetColumnQuantifiers, RemoveNameQuantifiers, dbTypeEnds, getDbLabel, getMermaidDiagramDb } from "./utils/sharedUtils";
 import { pluginVersion } from "./utils/constants";
+import { dbToOpenApi } from "./utils/nosqlUtils";
 
 declare const window: Customwindow;
 
@@ -65,58 +66,6 @@ Draw.loadPlugin(function(ui) {
         }
         sqlInputGenSQL.value = result;
     };
-    /**
-     * convert db to openapi
-     * @param db 
-     * @returns 
-     */
-    function dbToOpenApi(db: DbDefinition): PartialOpenApiSchema {
-        const result: PartialOpenApiSchema = {
-            openapi: "3.0.0",
-            info: {
-                // drawio file name?
-                title: "drawio nosql export",
-                version: pluginVersion,
-                "x-comment": "Generated by from drawio uml using plugin nosql"
-            },
-            paths: {},
-            components: {
-                schemas: {}
-            }
-        };
-        const schema: JSONSchema4 = {};
-        const entities = db.getEntities();
-        for (const key in entities) {
-            if (Object.prototype.hasOwnProperty.call(entities, key)) {
-                const entity = entities[key];
-                if(schema[key]){
-                    continue;
-                }
-                schema[key] = {
-                    type: "object",
-                    title: key,
-                    additionalProperties: false,
-                    properties: {}
-                };
-                for (let p = 0; p < entity.attributes.length; p++) {
-                    const attribute = entity.attributes[p];
-                    const propName = attribute.attributeName;
-                    if(!propName || schema[key].properties[propName]){
-                        continue;
-                    }
-                    const attType = attribute.attributeType?.split(" ") ?? [];
-                    const property: JSONSchema4 = {
-                        title: `${key}.${propName}`,
-                        nullable: attribute.attributeType?.includes("nullable") ?? false,
-                        type: (attType[0] ?? "string") as JSONSchema4TypeName 
-                    };
-                    schema[key].properties[attribute.attributeName!] = property;
-                }
-            }
-        }
-        result.components!.schemas = schema;
-        return result;
-    }
 
     mxUtils.br(divGenSQL);
