+++ tests/utils/sharedUtils.spec.ts
@@ -1,7 +1,23 @@
 import { ColumnQuantifiers } from "@funktechno/sqlsimpleparser/lib/types";
-import { GetColumnQuantifiers, dbTypeEnds, getDbLabel, removeHtml } from "../../src/utils/sharedUtils";
+import {
+  GetColumnQuantifiers,
+  dbTypeEnds,
+  getDbLabel,
+  getMermaidDiagramDb,
+  removeHtml,
+  GenerateDatabaseModel,
+} from "../../src/utils/sharedUtils";
 import { multiAssert } from "../helpers";
-import { TableAttribute } from "../../src/types/sql-plugin-types";
+import { DatabaseModelResult, TableAttribute } from "../../src/types/sql-plugin-types";
+import {
+  DbDefinition,
+  DbEntityAttributesDefinition,
+  DbEntityDefinition,
+  DbRelSpec,
+  DbRelationshipDefinition,
+} from "@funktechno/little-mermaid-2-the-sql/lib/src/types";
+
+import '../../src/types/drawio-types';
 
 describe("sharedUtils.ts", () => {
   it("dbTypeEnds", () => {
@@ -15,8 +31,8 @@ describe("sharedUtils.ts", () => {
       ts: { Start: "`", End: "`" },
       openapi: { Start: "`", End: "`" },
       sqlserver: { Start: "[", End: "]" },
-      sqlite: { Start: "\"", End: "\"" },
-      postgres: { Start: "\"", End: "\"" },
+      sqlite: { Start: '"', End: '"' },
+      postgres: { Start: '"', End: '"' },
     };
 
     for (const key in testTypes) {
@@ -37,8 +53,8 @@ describe("sharedUtils.ts", () => {
     let type_undefined: undefined;
     testTheory.push(() =>
       expect(GetColumnQuantifiers(type_undefined)).toEqual({
-        Start: "\"",
-        End: "\"",
+        Start: '"',
+        End: '"',
       })
     );
 
@@ -48,13 +64,34 @@ describe("sharedUtils.ts", () => {
     const expectedResult = "text only";
     const testData = `<span>${expectedResult}</span>`;
     expect(removeHtml(testData)).toBe("text only");
-     
-  })
-  it("getDbLabel", () => {
-    const columnQuantifiers = GetColumnQuantifiers("mysql");
-    const rowValue = "`LastName` varchar(255)"
-    const result = getDbLabel(rowValue, columnQuantifiers);
-    const expectedResult:TableAttribute = { attributeName: "LastName", attributeType: "varchar(255)" }
+  });
+  it("getMermaidDiagramDb", () => {
+    const mockDrawioUI: DrawioUI = {
+      fileNode: null,
+      hideDialog:() => {},
+      showDialog:(...args: any[]) => {},
+      editor: {
+        graph: {
+          getModel:() => {
+            const cells: Record<any, DrawioCell> = {};
+            return {cells};
+          }
+        } as any
+      },
+      actions: {
+        addAction:(name: string, action: () => void)=> {},
+        get:(name: string)=> { return {funct: () => {}} } 
+      },
+      menus: {
+        get:(name: string) => null,
+        funct: (...args: any[]) => {},
+        enabled: true,
+        addMenuItems:(menu: any, arg: any, arg2: any) => {}
+      } as any,
+      importLocalFile:(args: boolean) => {}
+    };
+    const result = getMermaidDiagramDb(mockDrawioUI, "mysql");
+    const expectedResult = GenerateDatabaseModel({}, [])
     expect(result).toEqual(expectedResult);
-  })
+  });
 });
